oracle_grants: added option for proxy_users
The parameter proxy_users could be used to add proxy connections.

This is the 1st step of the implementation. There is only support for adding
grant at the moment. The list with name: testproxy give us flexibility for
more options in the future.

proxy_users:
  - name: testproxy

alter user TESTUSER grant connect through testproxy;